Extract driver nodes reconciliation to its own file
